requirements.txt
# 核心数据分析库
pandas>=1.5.0
numpy>=1.24.0

# 量化交易框架
vectorbt>=0.25.0
backtrader>=1.9.78.123

# 数据处理和缓存
joblib>=1.2.0
scikit-learn>=1.2.0  # 用于base模块

# 数据存储格式
pyarrow>=11.0.0  # Parquet支持
fastparquet>=2023.4.0

# 可视化和绘图
plotly>=5.13.0
dash>=2.9.0
dash-bootstrap-components>=1.4.0

# 异步和网络请求
aiohttp>=3.8.0
websockets>=11.0.0

# 进度显示
tqdm>=4.65.0

# 开发工具
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0

# 测试框架
pytest>=7.3.0
pytest-asyncio>=0.21.0

# 其他工具
python-dotenv>=1.0.0
pyyaml>=6.0

# 可选：交易所API SDK
# ccxt>=4.0.0
# okex-python-sdk-api>=1.0.0